#10 - Service Layer CUD 기능 인증 기능 추가 수정
* Article/ArticleComment/UserAccount CommandService에 현재 인증 객체의 username을 parameter로 전달 받아 생성/수정/삭제 권한 확인하는 로직 추가 수정함.
* Article/ArticleComment CommandService의 대부분의 로직에서 연관관계에 있는 UserAccount 엔티티를 조회하기 때문에 통신되는 쿼리 성능 개선을 위해 각 엔티티 Repository에 Querydsl을 사용해 UserAccount를 페치조인하는 단건 조회 쿼리 구현함.
* 권한 체크시 발생할 수 있는 에러를 CustomException으로 구현해 등록함.
* 조회 권한 없는 경우 :  NoAuthorityToReadException
* 수정/삭제 권한 없는 경우 : NoAuthorityToUpdateDeleteException

+++ src/main/java/com/example/projectboard/application/articlecomments/ArticleCommentCommandServiceImpl.java
@@ -1,10 +1,13 @@
 package com.example.projectboard.application.articlecomments;
 
 import com.example.projectboard.common.exception.EntityNotFoundException;
+import com.example.projectboard.common.exception.NoAuthorityToUpdateDeleteException;
+import com.example.projectboard.common.exception.UsernameNotFoundException;
 import com.example.projectboard.domain.articlecomments.ArticleCommentCommand;
 import com.example.projectboard.domain.articlecomments.ArticleCommentInfo;
 import com.example.projectboard.domain.articlecomments.ArticleCommentRepository;
 import com.example.projectboard.domain.articles.ArticleRepository;
+import com.example.projectboard.domain.users.UserAccountRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,44 +20,86 @@ public class ArticleCommentCommandServiceImpl implements ArticleCommentCommandSe
 
     private final ArticleCommentRepository articleCommentRepository;
     private final ArticleRepository articleRepository;
+    private final UserAccountRepository userAccountRepository;
 
+    /**
+     * 새로운 ArticleComment 등록 메서드.
+     * @param username : 현재 인증 객체(Security Context)에 담겨 있는 사용자 정보에서의 username (String)
+     * @param command : ArticleComment 등록을 위한 데이터가 담긴 객체 (ArticleCommentCommand.RegisterReq)
+     * @return ArticleCommentInfo.MainInfo : 등록된 ArticleComment 엔티티를 ArticleCommentInfo.MainInfo에 감싸서 반환.
+     */
     @Override
     @Transactional
-    public ArticleCommentInfo.MainInfo registerComment(ArticleCommentCommand.RegisterReq command) {
-        log.info("{}:{}", getClass().getSimpleName(), "registerComment(ArticleCommentCommand.RegisterReq)");
+    public ArticleCommentInfo.MainInfo registerComment(String username, ArticleCommentCommand.RegisterReq command) {
+        log.info("{}:{}", getClass().getSimpleName(), "registerComment(String, ArticleCommentCommand.RegisterReq)");
 
-        // 게시글 엔티티 조회
+        // 게시글(Article) 엔티티 조회
         var article = articleRepository.findById(command.getArticleId())
                 .orElseThrow(EntityNotFoundException::new);
 
+        // 작성자(UserAccount) 엔티티 조회
+        var userAccount = userAccountRepository.findByUsername(username).orElseThrow(() ->
+                new UsernameNotFoundException("유저를 찾을 수 없습니다. username=" + username));
+
         // 댓글 등록
-        var savedComment = articleCommentRepository.save(command.toEntity(article));
+        var savedComment = articleCommentRepository.save(command.toEntity(article, userAccount));
         return new ArticleCommentInfo.MainInfo(savedComment, command.getArticleId());
     }
 
+    /**
+     * ArticleComment 내용 수정 메서드.
+     * @param commentId : 수정하려는 ArticleComment의 id (Long)
+     * @param username : 현재 인증 객체(Security Context)에 담겨 있는 사용자 정보에서의 username (String) -> 수정 요청 권한 확인을 위해
+     * @param command : ArticleComment 수정을 위한 데이터가 담긴 객체 (ArticleCommentCommand.UpdateReq)
+     */
     @Override
     @Transactional
-    public void update(Long commentId, ArticleCommentCommand.UpdateReq command) {
-        log.info("{}:{}", getClass().getSimpleName(), "update(Long, ArticleCommentCommand.UpdateReq)");
+    public void update(Long commentId, String username, ArticleCommentCommand.UpdateReq command) {
+        log.info("{}:{}", getClass().getSimpleName(), "update(Long, String, ArticleCommentCommand.UpdateReq)");
 
-        // 업데이트할 엔티티 조회
-        var comment = articleCommentRepository.findById(commentId)
+        // 업데이트할 ArticleComment 엔티티 조회
+        var comment = articleCommentRepository.findByIdFetchJoinUserAccount(commentId)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
 
-        // 댓글 수정
-        comment.update(command.getContent());
+        // 사용자 계정 엔티티 조회
+        var userAccount = userAccountRepository.findByUsername(username).orElseThrow(() ->
+                new UsernameNotFoundException("유저를 찾을 수 없습니다. username=" + username));
+
+        // 수정 권한 확인
+        if (comment.getUserAccount().equals(userAccount)) {
+            // 댓글 수정
+            comment.update(command.getContent());
+        } else {
+            log.error("수정 권한이 없는 사용자입니다. username={}", username);
+            throw new NoAuthorityToUpdateDeleteException();
+        }
     }
 
+    /**
+     * ArticleComment 삭제 메서드
+     * @param commentId : 삭제하려는 ArticleComment id (Long)
+     * @param username : 현재 인증 객체(Security Context)에 담겨 있는 사용자 정보에서의 username (String) -> 삭제 요청 권한 확인을 위해
+     */
     @Override
     @Transactional
-    public void delete(Long commentId) {
-        log.info("{}:{}", getClass().getSimpleName(), "delete(Long)");
+    public void delete(Long commentId, String username) {
+        log.info("{}:{}", getClass().getSimpleName(), "delete(Long, String)");
 
-        // 엔티티 조회
-        var comment = articleCommentRepository.findById(commentId)
+        // ArticleComment 엔티티 조회
+        var comment = articleCommentRepository.findByIdFetchJoinUserAccount(commentId)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 댓글입니다."));
 
-        // 댓글 삭제
-        articleCommentRepository.delete(comment);
+        // 사용자 계정 엔티티 조회
+        var userAccount = userAccountRepository.findByUsername(username).orElseThrow(() ->
+                new UsernameNotFoundException("유저를 찾을 수 없습니다. username=" + username));
+
+        // 삭제 권한 확인
+        if (comment.getUserAccount().equals(userAccount)) {
+            // 댓글 삭제
+            articleCommentRepository.delete(comment);
+        } else {
+            log.error("삭제 권한이 없는 사용자입니다. username={}", username);
+            throw new NoAuthorityToUpdateDeleteException();
+        }
     }
 }
